Projects/NewsLetter/requirements.txt
google-genai
prefect
prefect[docker]
python-dotenv
premailer